requirements.txt
dhooks
keyboard
pyperclip
python-dotenv